Fixed row highlighting

+++ src/main/java/org/oktanauts/model/Measurement.java
@@ -6,13 +6,13 @@ public class Measurement {
 
     private String code;
     private String name;
-    private float value;
+    private double value;
     private String unit;
     private Timestamp timestamp;
     private Patient patient;
 
 
-    public Measurement(String code, String name, Float value, String unit, Timestamp timestamp, Patient p) {
+    public Measurement(String code, String name, double value, String unit, Timestamp timestamp, Patient p) {
         this.code = code;
         this.name = name;
         this.value = value;
@@ -21,13 +21,11 @@ public class Measurement {
         this.patient = p;
     }
 
-
-
     public Patient getPatient() {
         return patient;
     }
 
-    public float getValue() { return this.value; }
+    public double getValue() { return this.value; }
 
     public Timestamp getTimestamp() { return this.timestamp; }
